Skapa egen klass till mainmenu's bakgrund
